feat:myActivity.ts 파일에 병합-#116
