+++ omechu-app/src/app/mypage/api/myActivity.ts
@@ -1,6 +1,26 @@
 import apiClient from "@/lib/api/client";
 
-export const fetchRestaurants = async (userId: string) => {
-  const res = await apiClient.get(`/test/restaurants/${userId}`);
-  return res.data.success?.data ?? [];
-};
+export interface RestaurantItem {
+  id: number;
+  name: string;
+  address?: string;
+  rating?: number;
+  // images?: string[]; // 추후 API에서 제공되면 반영
+}
+
+export interface FetchRestaurantsResponse {
+  resultType: string;
+  error: null | any;
+  success: {
+    data: RestaurantItem[];
+  };
+}
+
+export async function fetchRestaurants(
+  userId: string,
+): Promise<FetchRestaurantsResponse> {
+  const res = await apiClient.get<FetchRestaurantsResponse>(
+    `/restaurants/${userId}`,
+  );
+  return res.data;
+}
